edit api user
