+++ src/boot/firebase.js
@@ -3,13 +3,13 @@ import Firebase from 'firebase/app'
 export default ({ Vue }) => {
 
     const FirebaseConfig = {
-        apiKey: "AIzaSyCYsY6bdPLNl-0UjCJ-RxJ_m5mpia46Nt8",
-        authDomain: "laberu-77a42.firebaseapp.com",
-        projectId: "laberu-77a42",
-        storageBucket: "laberu-77a42.appspot.com",
-        messagingSenderId: "993128334521",
-        appId: "1:993128334521:web:47b472ca44d8e6581dc768",
-        measurementId: "G-055DEPHQ9W"
+        apiKey: "AIzaSyCX5ErvKR5qhKdgeXAGHn5yEJdAGJfUpyI",
+        authDomain: "laberu.firebaseapp.com",
+        projectId: "laberu",
+        storageBucket: "laberu.appspot.com",
+        messagingSenderId: "887970080889",
+        appId: "1:887970080889:web:fe9bfb938753ef14568ca8",
+        measurementId: "G-9V6TZG3HVT"
     }
     Firebase.initializeApp(FirebaseConfig);
     Vue.prototype.$firebase = Firebase;
